C++ API: (LEC) move structs LEC_CONST_REAL, LEC_CONST_LONG, LEC_COEF, LEC_VAR, LEC_PERIOD to new header file l_value.h
